fix bug in SimplePageRankRunner when converting long to float

+++ src/SimplePageRank/PageRankRunner.java
@@ -37,7 +37,7 @@ public class PageRankRunner {
             FileOutputFormat.setOutputPath(job, new Path(outputPath + Conf.FILE_NAME + (i + 1)));
 
             job.waitForCompletion(true);
-            float residual = job.getCounters().findCounter(SimplePageRank.Counter.RESIDUAL_COUNTER).getValue() / Conf.MULTIPLE;
+            float residual = ((float) job.getCounters().findCounter(SimplePageRank.Counter.RESIDUAL_COUNTER).getValue()) / Conf.MULTIPLE;
             float avgError = residual / Conf.NODES_NUM;
 
             float threshold = Conf.EPSILON;
